pyproject.toml
[project]
name = "mecano-man"
version = "0.1.0"
description = "Multi-tenant RAG SaaS backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # FastAPI stack
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.12",
    
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "greenlet>=3.1.0",
    
    # Task queue
    "arq>=0.26.1",
    "redis>=5.2.0",
    
    # Vector store
    "qdrant-client>=1.12.0",
    
    # Object storage
    "boto3>=1.35.0",
    "aioboto3>=13.2.0",
    
    # LLM & Embeddings
    "openai>=1.57.0",
    "tiktoken>=0.8.0",
    
    # Document parsing
    "pypdf>=5.1.0",
    "python-docx>=1.1.0",
    "python-pptx>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",
    "markdown>=3.7",
    
    # Utils
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    
    # Authentication & Billing
    "clerk-backend-api>=1.0.0",
    "stripe>=8.0.0",
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",

    # PDF export
    "playwright>=1.49.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "ipython>=8.29.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
